🚚 Move ConsoleLogger to wc-gateway module
